fix: address cache hit rate review feedback (#90)
- sdk: default cache_hit_rate on deserialize so summaries serialized
  before the field existed still load (None = unavailable).
- sources: expose has_cache_read in table/CSV/JSON output so users can
  see why a source reports cache hit rate as N/A. Renderers move to
  sources_cmd.rs to keep app.rs under the module size limit.

Confidence: high
Scope-risk: narrow
Tested: cargo test (584 passed, +1 new), cargo clippy -D warnings, cargo fmt --check, checks/check_workflow.py, manual ccstats sources table/CSV output
Related: #90
