+++ src/app.rs
@@ -19,11 +19,10 @@ use crate::output::{
 };
 use crate::pricing::{CostDisplayMode, PricingDb};
 use crate::source::{
-    ALL_SOURCES, Capabilities, Source, all_capabilities, all_sources, load_blocks, load_daily,
-    load_projects, load_sessions, load_tool_calls,
+    Capabilities, Source, all_capabilities, all_sources, load_blocks, load_daily, load_projects,
+    load_sessions, load_tool_calls,
 };
 use crate::utils::{Timezone, filter_json};
-use serde_json::json;
 
 /// Print JSON output, optionally filtering through jq
 pub(crate) fn print_json(json: &str, jq_filter: Option<&str>) {
@@ -348,118 +347,6 @@ fn render_tools(summary: &ToolSummary, ctx: &CommandContext<'_>) {
     }
 }
 
-fn handle_sources(ctx: &CommandContext<'_>) {
-    let sources: Vec<&dyn Source> = all_sources().collect();
-    let mut all_caps = all_capabilities();
-    all_caps.has_projects = false;
-    all_caps.has_billing_blocks = false;
-
-    render_sources(&sources, &all_caps, ctx);
-}
-
-fn render_sources(sources: &[&dyn Source], all_caps: &Capabilities, ctx: &CommandContext<'_>) {
-    match ctx.cli.output_format() {
-        OutputFormat::Csv => render_sources_csv(sources, all_caps),
-        OutputFormat::Json => render_sources_json(sources, all_caps, ctx),
-        OutputFormat::Table => print_sources_table(sources, all_caps),
-    }
-}
-
-fn render_sources_csv(sources: &[&dyn Source], all_caps: &Capabilities) {
-    println!(
-        "name,display_name,aliases,has_projects,has_billing_blocks,has_reasoning_tokens,has_cache_creation,needs_dedup"
-    );
-    println!(
-        "{},All Sources,,{},{},{},{},{}",
-        ALL_SOURCES,
-        all_caps.has_projects,
-        all_caps.has_billing_blocks,
-        all_caps.has_reasoning_tokens,
-        all_caps.has_cache_creation,
-        all_caps.needs_dedup
-    );
-    for source in sources {
-        let caps = source.capabilities();
-        let aliases = source.aliases().join("|");
-        println!(
-            "{},{},{},{},{},{},{},{}",
-            source.name(),
-            source.display_name(),
-            aliases,
-            caps.has_projects,
-            caps.has_billing_blocks,
-            caps.has_reasoning_tokens,
-            caps.has_cache_creation,
-            caps.needs_dedup
-        );
-    }
-}
-
-fn render_sources_json(sources: &[&dyn Source], all_caps: &Capabilities, ctx: &CommandContext<'_>) {
-    let mut payload = vec![json!({
-        "name": ALL_SOURCES,
-        "display_name": "All Sources",
-        "aliases": [],
-        "capabilities": {
-            "has_projects": all_caps.has_projects,
-            "has_billing_blocks": all_caps.has_billing_blocks,
-            "has_reasoning_tokens": all_caps.has_reasoning_tokens,
-            "has_cache_creation": all_caps.has_cache_creation,
-            "needs_dedup": all_caps.needs_dedup
-        }
-    })];
-    payload.extend(sources.iter().map(|source| {
-        let caps = source.capabilities();
-        json!({
-            "name": source.name(),
-            "display_name": source.display_name(),
-            "aliases": source.aliases(),
-            "capabilities": {
-                "has_projects": caps.has_projects,
-                "has_billing_blocks": caps.has_billing_blocks,
-                "has_reasoning_tokens": caps.has_reasoning_tokens,
-                "has_cache_creation": caps.has_cache_creation,
-                "needs_dedup": caps.needs_dedup
-            }
-        })
-    }));
-    let json = serde_json::to_string(&payload).unwrap_or_else(|_| "[]".to_string());
-    print_json(&json, ctx.jq_filter);
-}
-
-fn print_sources_table(sources: &[&dyn Source], all_caps: &Capabilities) {
-    println!("Available sources:");
-    println!(
-        "- {} (All Sources) aliases: - | has_projects={} has_billing_blocks={} has_reasoning_tokens={} has_cache_creation={} needs_dedup={}",
-        ALL_SOURCES,
-        all_caps.has_projects,
-        all_caps.has_billing_blocks,
-        all_caps.has_reasoning_tokens,
-        all_caps.has_cache_creation,
-        all_caps.needs_dedup
-    );
-    for source in sources {
-        let caps = source.capabilities();
-        let aliases = if source.aliases().is_empty() {
-            "-".to_string()
-        } else {
-            source.aliases().join(", ")
-        };
-        println!(
-            "- {} ({}) aliases: {} | has_projects={} has_billing_blocks={} has_reasoning_tokens={} has_cache_creation={} needs_dedup={}",
-            source.name(),
-            source.display_name(),
-            aliases,
-            caps.has_projects,
-            caps.has_billing_blocks,
-            caps.has_reasoning_tokens,
-            caps.has_cache_creation,
-            caps.needs_dedup
-        );
-    }
-    println!("Hint: use `--source <name|alias>` (e.g. `--source codex` or `--source cx`).");
-}
-
 /// Statusline keeps its compact single-line semantics outside generic output dispatch.
 fn handle_statusline(source: &dyn Source, ctx: &CommandContext<'_>) {
     let result = load_daily(source, ctx.filter, ctx.timezone, true, false);
@@ -632,7 +519,7 @@ pub(crate) fn handle_source_command(
     let caps = source.capabilities();
 
     match command {
-        SourceCommand::Sources => return handle_sources(ctx),
+        SourceCommand::Sources => return crate::sources_cmd::handle_sources(ctx),
         SourceCommand::Session => return handle_session(source, ctx),
         SourceCommand::Project => {
             if !caps.has_projects {
@@ -705,7 +592,7 @@ fn load_all_daily(ctx: &CommandContext<'_>, quiet: bool) -> (LoadResult, Capabil
 /// Handle aggregate commands across every registered data source.
 pub(crate) fn handle_all_sources_command(command: SourceCommand, ctx: &CommandContext<'_>) {
     match command {
-        SourceCommand::Sources => return handle_sources(ctx),
+        SourceCommand::Sources => return crate::sources_cmd::handle_sources(ctx),
         SourceCommand::Statusline => {
             let (result, caps) = load_all_daily(ctx, true);
             if ctx.cli.json {
